fix to the foreign key error on delete project

+++ api.php
@@ -51,7 +51,7 @@ class timesheetApi{
 	}
 	public function addNewProject($post){
 		//validate 
-		if(empty($post['name']) || empty($post['description']))
+		if(empty($post['name']))
 			return false;
 		//create query
 		$query = 'INSERT INTO `project` 
@@ -70,7 +70,18 @@ class timesheetApi{
 		if(empty($post['id']) || !preg_match("/project\-[0-9]+/",$post['id']))
 			return false;
 		$post['id'] = str_replace('project-','',$post['id']);
-		//create query
+		//delete all the worklogs
+		$query = 'DELETE FROM `work_log`	WHERE `project_id` = "'.mysql_escape_string($post['id']).'" ';
+		if(!$this->connection->query($query)){
+			$this->error = $this->connection->getError();
+			return false;
+		}
+		$query = 'DELETE FROM `task`	WHERE `project_id` = "'.mysql_escape_string($post['id']).'" ';
+		if(!$this->connection->query($query)){
+			$this->error = $this->connection->getError();
+			return false;
+		}
+		//delete the project
 		$query = 'DELETE FROM `project`	WHERE `id` = "'.mysql_escape_string($post['id']).'" ';
 		//execute
 		if(!$this->connection->query($query)){
